Move some repository methods from models to modules/repository
